upgrade to dioxus v0.7.8
