+++ Cargo.toml
@@ -6,8 +6,10 @@ members = ["primitives", "preview", "test-harness", "dioxus-attributes"]
 dioxus-primitives = { path = "primitives" }
 tracing = { version = "0.1", features = ["std"] }
 
-# dioxus = "0.7.7"
-dioxus = { path = "../../dioxus/packages/dioxus" }
+dioxus = "0.7.8"
+dioxus-rsx = "0.7.8"
+# dioxus = { path = "../../dioxus/packages/dioxus" }
+# dioxus-rsx =  {path = "../../dioxus/packages/rsx" }
 
 [profile.release]
 opt-level = "z"
@@ -16,3 +18,10 @@ lto = true
 codegen-units = 1
 strip = true
 incremental = false
+
+[profile.dev]
+debug = 0
+
+[profile.dev.build-override]
+opt-level = 3
+codegen-units = 1
