+++ dioxus-attributes/Cargo.toml
@@ -9,7 +9,7 @@ license = "MIT OR Apache-2.0"
 proc-macro = true
 
 [dependencies]
-dioxus-rsx = "0.7.1"
+dioxus-rsx = {  workspace = true }
 proc-macro2 = "1.0"
 quote = "1.0"
 syn = { version = "2.0", features = ["full", "extra-traits"] }
