+++ preview/src/dashboard/views/email_client/state.rs
@@ -52,7 +52,7 @@ impl EmailClientState {
 
 #[store(pub)]
 impl<Lens> Store<EmailClientState, Lens> {
-    pub fn active_folder_label(&self) -> &'static str {
+    fn active_folder_label(&self) -> &'static str {
         let active_folder = self.active_folder().cloned();
         FOLDERS
             .iter()
@@ -61,7 +61,7 @@ impl<Lens> Store<EmailClientState, Lens> {
             .unwrap_or("Inbox")
     }
 
-    pub fn visible_message_ids(&self) -> Vec<String> {
+    fn visible_message_ids(&self) -> Vec<String> {
         let active_folder = self.active_folder().cloned();
         let active_tab = self.active_tab().cloned();
         let search_query = self.search_query().cloned();
@@ -88,7 +88,7 @@ impl<Lens> Store<EmailClientState, Lens> {
             .collect()
     }
 
-    pub fn selected_message_uid(&self, visible_ids: &[String]) -> Option<String> {
+    fn selected_message_uid(&self, visible_ids: &[String]) -> Option<String> {
         let selected_id = self.selected_id().cloned();
         if visible_ids.iter().any(|uid| uid == &selected_id) {
             return Some(selected_id);
@@ -97,7 +97,7 @@ impl<Lens> Store<EmailClientState, Lens> {
         visible_ids.first().cloned()
     }
 
-    pub fn selected_message_index(
+    fn selected_message_index(
         &self,
         selected_uid: Option<&str>,
         visible_ids: &[String],
@@ -112,7 +112,7 @@ impl<Lens> Store<EmailClientState, Lens> {
             .unwrap_or(0)
     }
 
-    pub fn folder_count(&self, folder_id: FolderId) -> u32 {
+    fn folder_count(&self, folder_id: FolderId) -> u32 {
         let messages_store = self.messages();
         let count = messages_store
             .read()
@@ -122,7 +122,7 @@ impl<Lens> Store<EmailClientState, Lens> {
         count
     }
 
-    pub fn tab_count(
+    fn tab_count(
         &self,
         tab_id: TabId,
         visible_query: &str,
@@ -146,63 +146,63 @@ impl<Lens> Store<EmailClientState, Lens> {
         count
     }
 
-    pub fn set_active_folder(&mut self, folder_id: FolderId) {
+    fn set_active_folder(&mut self, folder_id: FolderId) {
         self.active_folder().set(folder_id);
         self.close_read_pane();
     }
 
-    pub fn set_active_tab(&mut self, tab_id: TabId) {
+    fn set_active_tab(&mut self, tab_id: TabId) {
         self.active_tab().set(tab_id);
         self.close_read_pane();
     }
 
-    pub fn set_search_query(&mut self, query: String) {
+    fn set_search_query(&mut self, query: String) {
         self.search_query().set(query);
         self.close_read_pane();
     }
 
-    pub fn set_selected_tags(&mut self, tags: Vec<MessageTag>) {
+    fn set_selected_tags(&mut self, tags: Vec<MessageTag>) {
         self.selected_tags().set(tags);
         self.close_read_pane();
     }
 
-    pub fn select_message(&mut self, uid: String) {
+    fn select_message(&mut self, uid: String) {
         self.selected_id().set(uid);
         self.read_open().set(true);
     }
 
-    pub fn close_read_pane(&mut self) {
+    fn close_read_pane(&mut self) {
         self.read_open().set(false);
     }
 
-    pub fn open_compose(&mut self) {
+    fn open_compose(&mut self) {
         self.compose_open().set(true);
     }
 
-    pub fn set_compose_open(&mut self, open: bool) {
+    fn set_compose_open(&mut self, open: bool) {
         self.compose_open().set(open);
     }
 
-    pub fn discard_compose(&mut self) {
+    fn discard_compose(&mut self) {
         self.compose_open().set(false);
         self.compose_to().set(String::new());
         self.compose_subject().set(String::new());
         self.compose_body().set(String::new());
     }
 
-    pub fn set_compose_to(&mut self, value: String) {
+    fn set_compose_to(&mut self, value: String) {
         self.compose_to().set(value);
     }
 
-    pub fn set_compose_subject(&mut self, value: String) {
+    fn set_compose_subject(&mut self, value: String) {
         self.compose_subject().set(value);
     }
 
-    pub fn set_compose_body(&mut self, value: String) {
+    fn set_compose_body(&mut self, value: String) {
         self.compose_body().set(value);
     }
 
-    pub fn archive_message(&mut self, uid: String) {
+    fn archive_message(&mut self, uid: String) {
         self.update_message(uid.clone(), |message| {
             message.folder_id = FolderId::Archive;
             message.unread = false;
@@ -210,14 +210,14 @@ impl<Lens> Store<EmailClientState, Lens> {
         self.close_if_selected(&uid);
     }
 
-    pub fn snooze_message(&mut self, uid: String) {
+    fn snooze_message(&mut self, uid: String) {
         self.update_message(uid.clone(), |message| {
             message.snoozed = true;
         });
         self.close_if_selected(&uid);
     }
 
-    pub fn delete_message(&mut self, uid: String) {
+    fn delete_message(&mut self, uid: String) {
         self.update_message(uid.clone(), |message| {
             message.folder_id = FolderId::Trash;
             message.unread = false;
@@ -225,45 +225,45 @@ impl<Lens> Store<EmailClientState, Lens> {
         self.close_if_selected(&uid);
     }
 
-    pub fn toggle_message_flag(&mut self, uid: String) {
+    fn toggle_message_flag(&mut self, uid: String) {
         self.update_message(uid, |message| {
             message.flagged = !message.flagged;
         });
     }
 
-    pub fn toggle_message_star(&mut self, uid: String) {
+    fn toggle_message_star(&mut self, uid: String) {
         self.update_message(uid, |message| {
             message.starred = !message.starred;
         });
     }
 
-    pub fn move_message_to_trash(&mut self, uid: String) {
+    fn move_message_to_trash(&mut self, uid: String) {
         self.update_message(uid.clone(), |message| {
             message.folder_id = FolderId::Trash;
         });
         self.close_if_selected(&uid);
     }
 
-    pub fn remove_message_tag(&mut self, uid: String, tag: MessageTag) {
+    fn remove_message_tag(&mut self, uid: String, tag: MessageTag) {
         self.update_message(uid, |message| {
             message.tags.retain(|current| *current != tag);
         });
     }
 
-    pub fn set_message_tags(&mut self, uid: String, tags: Vec<MessageTag>) {
+    fn set_message_tags(&mut self, uid: String, tags: Vec<MessageTag>) {
         self.update_message(uid, |message| {
             message.tags = tags;
         });
     }
 
-    pub fn update_message(&mut self, uid: String, update: impl FnOnce(&mut MessageState)) {
+    fn update_message(&mut self, uid: String, update: impl FnOnce(&mut MessageState)) {
         if let Some(mut message) = self.messages().get(uid) {
             let mut message = message.write();
             update(&mut message);
         }
     }
 
-    pub fn close_if_selected(&mut self, uid: &str) {
+    fn close_if_selected(&mut self, uid: &str) {
         if self.selected_id().cloned() == uid {
             self.close_read_pane();
         }
